file DNE error fix

+++ master_detect_data.py
@@ -65,8 +65,9 @@ if __name__ == "__main__":
           folders.append(line.strip())
     os.remove(yolo_output_path_log)
     for i in range(len(folders)):
-        vid = os.path.join(folders[i], os.listdir(folders[i])[0])
-        shutil.move(vid, full_output_path)
-        os.rmdir(folders[i])
+        if len(folders[i] > 0):
+          vid = os.path.join(folders[i], os.listdir(folders[i])[0])
+          shutil.move(vid, full_output_path)
+          os.rmdir(folders[i])
   if os.path.exists("runs"):
     shutil.rmtree("runs")
